automatically backed up by learn

+++ lib/create.sql
@@ -15,7 +15,7 @@ age INTEGER
 
 CREATE TABLE pledges (
 id INTEGER PRIMARY KEY,
-amount INTEGER,
+amount REAL,
 user_id INTEGER,
 project_id INTEGER
 );
